Rename object->data
